refactor: FileUploader 인터페이스 추상화 적용
